+++ DTCD-StyleSystem/src/StyleSystem.js
@@ -1,4 +1,12 @@
-import { InteractionSystemAdapter, EventSystemAdapter, SystemPlugin } from '../../DTCD-SDK/index';
+import {
+  InteractionSystemAdapter,
+  EventSystemAdapter,
+  SystemPlugin,
+  LogSystemAdapter,
+} from '../../DTCD-SDK/index';
+
+import './../fonts/fonts.css';
+import baseComponentList from './base-components/components';
 
 export class StyleSystem extends SystemPlugin {
   static getRegistrationMeta() {
@@ -6,6 +14,9 @@ export class StyleSystem extends SystemPlugin {
       name: 'StyleSystem',
       type: 'core',
       title: 'Дизайн система',
+      version: '0.2.0',
+      priority: 3,
+      withDependencies: false,
     };
   }
 
@@ -14,45 +25,78 @@ export class StyleSystem extends SystemPlugin {
     this.guid = guid;
     this.interactionSystem = new InteractionSystemAdapter();
     this.eventSystem = new EventSystemAdapter();
-    this.currentThemeName = 'light';
+    this.logSystem = new LogSystemAdapter(guid, 'StyleSystem');
+  }
+
+  async init() {
+    this.logSystem.info('Initializing system');
+
+    baseComponentList.forEach(component => {
+      const { name, baseClass } = component;
+      baseClass.guid = this.guid;
+      window.customElements.define(name, baseClass);
+    });
+
+    try {
+      this.logSystem.debug('Requesting design object from endpoint /get-design-objects');
+      const { data } = await this.interactionSystem.GETRequest(
+        '/mock_server/v1/get-design-objects'
+      );
+      this.logSystem.debug('Setting themes received from server in system');
+      this.themes = data;
+      this.logSystem.debug(`Setting ${this.themes[0].name} as default theme in system`);
+      this.currentThemeName = this.themes[0].name;
+      this.logSystem.info('System inited successfully');
+    } catch (err) {
+      this.logSystem.fatal(
+        `'${err.name}' occured while fetching desing object.${err.message}, ${err.stack}`
+      );
+    }
   }
 
   setTheme(name) {
+    this.logSystem.debug(`setTheme method called with argument ${name}`);
     if (typeof name === 'string') {
       const theme = this.themes.find(theme => theme.name === name);
       if (theme) {
         this.currentThemeName = name;
+        this.logSystem.info(`New theme '${name}' set in system`);
         this.eventSystem.createAndPublish(this.guid, 'ThemeUpdate');
       } else {
-        console.error(`Theme: ${name} doesn't exist`);
+        this.logSystem.warn(`Theme '${name}' doesn't exist in system!`);
         throw new Error('Theme not found!');
       }
     } else {
+      this.logSystem.debug(`argument type of '${name}' is not string `);
       throw new Error('Wrong argument type!');
     }
   }
 
-  async getThemes() {
-    const { data } = await this.interactionSystem.GETRequest('/get-design-objects');
-    this.themes = data;
-    return data;
+  getThemes() {
+    this.logSystem.info(`Returning all themes stored in system`);
+    return this.themes;
   }
 
   getCurrentTheme() {
+    this.logSystem.info(`Returning current theme of application`);
     return this.themes.find(theme => theme.name === this.currentThemeName);
   }
 
   setVariablesToElement(element, obj, startPrefix = '-') {
-    function setVariables(obj, prefix) {
+    const setVariables = (obj, prefix) => {
       Object.keys(obj).forEach(key => {
         if (typeof obj[key] === 'object') {
           const newPrefix = `${prefix}-${key}`;
           setVariables(obj[key], newPrefix);
         } else {
+          this.logSystem.debug(`setting '${prefix}-${key}' variable value to '${obj[key]}'`);
           element.style.setProperty(`${prefix}-${key}`, obj[key]);
         }
       });
-    }
+    };
     setVariables(obj.styleVariables, startPrefix);
+    this.logSystem.info(
+      `CSS variables from theme '${this.currentThemeName}' are installed for element ${element}`
+    );
   }
 }