+++ CHANGELOG.md
@@ -5,6 +5,35 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0]
+
+### Added
+
+- init hook to get themes from server
+- manifest file for dependencies
+- BaseButton custom component
+- Proxima Nova font
+- BaseTextarea custom component
+- BaseCheckbox custom component
+- BaseSwitch custom component
+- BaseColorPicker custom
+- BaseInput value setter and getter
+- DataSourceSelect
+- GaugeSegmentBuilder
+
+### Changed
+
+- _getThemes_ method just returns theme object and no longer requesys themes from server
+- deployment section in [README.md](Readme.md)
+- SDK version to 0116
+
+### Fixed
+
+- rollup build path for development
+- setVariables function in setVariablesToElement method
+- rollup building directory path
+- BaseCheckbox icon
+
 ## [0.2.0] - 2021-02-11
 
 ### Changed